backend/pyproject.toml
[project]
name = "product-intelligence-backend"
version = "1.0.0"
description = "Django backend for Product Intelligence Platform — multi-tenant AI-native SaaS"
requires-python = ">=3.11"

dependencies = [
    "django>=4.2,<5.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.1",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.1",
    "django-cors-headers>=4.3.1",
    "python-dotenv>=1.0.0",
    "celery>=5.3.4",
    "django-celery-beat>=2.6.0",
    "django-celery-results>=2.5.1",
    "django-filter>=23.5",
    "black>=24.1.1",
    "httpx>=0.25.2",
    "flake8>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-django>=4.7.0",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=4.1.0",
    "coverage>=7.3.2",
    "factory-boy>=3.3.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
python_files = "tests.py test_*.py *_test.py"
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py311"
